+++ src/commands/referral/cleanupExpiredCode.ts
@@ -1,41 +1,27 @@
-import { ReferralCode } from '@prisma/client';
 import { getPrismaClient } from '../../clients';
-import { parseDate } from './parseDate';
+import { OpResult } from '../../utils/opResult';
 
-export const cleanupExpiredCode = (
-  referrals: ReferralCode[]
-): ReferralCode[] => {
+export const cleanupExpiredCode = async (): Promise<OpResult<unknown>> => {
   const prisma = getPrismaClient();
+  const currentDate = new Date();
+  try {
+    const result = await prisma.referralCode.deleteMany({
+      where: {
+        expiry_date: {
+          lt: currentDate,
+        },
+      },
+    });
 
-  const [expiredIds, filteredReferrals] = referrals.reduce<
-    [string[], ReferralCode[]]
-  >(
-    (total, referral) => {
-      const expiryDate = new Date(referral.expiry_date).toLocaleDateString(
-        'en-AU'
-      );
-      const [parseDateCase] = parseDate(expiryDate);
-      if (parseDateCase !== 'SUCCESS') {
-        total[0].push(referral.id);
-      } else {
-        total[1].push(referral);
-      }
-
-      return total;
-    },
-    [[], []]
-  );
-
-  // SOMETIMES try to clean expired referral codes
-  if (expiredIds.length > 10) {
-    try {
-      prisma.referralCode.deleteMany({
-        where: { id: { in: expiredIds } },
-      });
-    } catch (error) {
-      console.error(error);
-    }
+    return {
+      success: true,
+      data: result,
+    };
+  } catch (error) {
+    console.error(error);
+    return {
+      success: false,
+      error,
+    };
   }
-
-  return filteredReferrals;
 };